+++ tests/test_data/test_datasets/test_ava_dataset.py
@@ -14,6 +14,7 @@ class TestAVADataset:
     def setup_class(cls):
         cls.data_prefix = osp.normpath(
             osp.join(osp.dirname(__file__), '../../data', 'ava_dataset'))
+        cls.label_file = osp.join(cls.data_prefix, 'action_list.txt')
         cls.ann_file = osp.join(cls.data_prefix, 'ava_sample.csv')
         cls.exclude_file = osp.join(cls.data_prefix,
                                     'ava_excluded_timestamps_sample.csv')
@@ -54,6 +55,26 @@ class TestAVADataset:
         assert len(ava_infos[0]['ann']) == 3
         target_labels = np.array([12, 17, 79])
         labels = np.zeros([81])
+        labels[target_labels] = 1.
+        target_labels = labels[None, ...]
+        assert_array_equal(ava_infos[0]['ann']['gt_labels'], target_labels)
+        assert_array_equal(ava_infos[0]['ann']['gt_bboxes'],
+                           np.array([[0.031, 0.162, 0.67, 0.995]]))
+        assert_array_equal(ava_infos[0]['ann']['entity_ids'], np.array([0]))
+
+        # custom classes
+        ava_dataset = AVADataset(
+            self.ann_file,
+            self.exclude_file,
+            self.pipeline,
+            label_file=self.label_file,
+            custom_classes=[17, 79],
+            num_classes=3,
+            data_prefix=self.data_prefix,
+            proposal_file=self.proposal_file)
+        ava_infos = ava_dataset.video_infos
+        target_labels = np.array([1, 2])
+        labels = np.zeros([3])
         labels[target_labels] = 1.
         target_labels = labels[None, ...]
         assert_array_equal(ava_infos[0]['ann']['gt_labels'], target_labels)
@@ -139,6 +160,7 @@ class TestAVADataset:
             osp.join(osp.dirname(__file__), '../../data', 'eval_detection'))
         ann_file = osp.join(data_prefix, 'gt.csv')
         label_file = osp.join(data_prefix, 'action_list.txt')
+
         ava_dataset = AVADataset(
             ann_file, None, [], label_file=label_file, num_classes=4)
         fake_result = [[
@@ -167,3 +189,31 @@ class TestAVADataset:
                        ]]
         res = ava_dataset.evaluate(fake_result)
         assert_array_almost_equal(res['mAP@0.5IOU'], 0.027777778)
+
+        # custom classes
+        ava_dataset = AVADataset(
+            ann_file,
+            None, [],
+            label_file=label_file,
+            num_classes=3,
+            custom_classes=[1, 3])
+        fake_result = [[
+            np.array([[0.362, 0.156, 0.969, 0.666, 0.106],
+                      [0.442, 0.083, 0.721, 0.947, 0.162]]),
+            np.array([[0.417, 0.167, 0.843, 0.939, 0.015],
+                      [0.35, 0.421, 0.57, 0.689, 0.427]])
+        ],
+                       [
+                           np.array([[0.256, 0.338, 0.726, 0.799, 0.563],
+                                     [0.071, 0.256, 0.64, 0.75, 0.297]]),
+                           np.array([[0.051, 0.005, 0.975, 0.942, 0.424],
+                                     [0.347, 0.05, 0.97, 0.944, 0.396]])
+                       ],
+                       [
+                           np.array([[0.39, 0.087, 0.833, 0.616, 0.447],
+                                     [0.461, 0.212, 0.627, 0.527, 0.036]]),
+                           np.array([[0.206, 0.456, 0.564, 0.725, 0.685],
+                                     [0.106, 0.445, 0.782, 0.673, 0.367]])
+                       ]]
+        res = ava_dataset.evaluate(fake_result)
+        assert_array_almost_equal(res['mAP@0.5IOU'], 0.04166667)